new gSDK

+++ src/mavlink/include/mavlink/v2.0/mavlink_conversions.h
@@ -38,10 +38,10 @@
  */
 MAVLINK_HELPER void mavlink_quaternion_to_dcm(const float quaternion[4], float dcm[3][3])
 {
-    double a = quaternion[0];
-    double b = quaternion[1];
-    double c = quaternion[2];
-    double d = quaternion[3];
+    double a = (double)quaternion[0];
+    double b = (double)quaternion[1];
+    double c = (double)quaternion[2];
+    double d = (double)quaternion[3];
     double aSq = a * a;
     double bSq = b * b;
     double cSq = c * c;
@@ -72,19 +72,25 @@ MAVLINK_HELPER void mavlink_dcm_to_euler(const float dcm[3][3], float* roll, flo
     theta = asinf(-dcm[2][0]);
 
     if (fabsf(theta - (float)M_PI_2) < 1.0e-3f) {
+
         phi = 0.0f;
         psi = (atan2f(dcm[1][2] - dcm[0][1],
                 dcm[0][2] + dcm[1][1]) + phi);
 
-    } else if (fabsf(theta + (float)M_PI_2) < 1.0e-3f) {
+    } 
+    else if (fabsf(theta + (float)M_PI_2) < 1.0e-3f) {
+
         phi = 0.0f;
         psi = atan2f(dcm[1][2] - dcm[0][1],
                   dcm[0][2] + dcm[1][1] - phi);
 
-    } else {
+    } 
+    else {
+
         phi = atan2f(dcm[2][1], dcm[2][2]);
         psi = atan2f(dcm[1][0], dcm[0][0]);
     }
+    // DatNguyen test
 
     *roll = phi;
     *pitch = theta;
